ultumos ajustes na criacao de usuario
