+++ src/components/form/FormInputSelect.tsx
@@ -51,7 +51,7 @@ export default function FormInputSelect({
       {label && (
         <label
           htmlFor={id || "teste"}
-          className="text-gray-500 dark:text-gray-400"
+          className="text-gray-500 dark:text-gray-400 "
         >
           {label}
         </label>
@@ -61,11 +61,11 @@ export default function FormInputSelect({
         id={id}
         onChange={handleChange}
         value={value}
-        className="focus-within:border-primary-600 hover:border-primary-600 shadow-primary-300/40 flex w-full cursor-pointer items-center justify-center gap-2 rounded-[12px] border-2 border-transparent bg-gray-400/30 py-2 pl-4 text-xl outline-0 transition-all outline-none placeholder:text-xl placeholder:text-gray-600/50 focus-within:bg-gray-400/30 focus-within:shadow-[0px_0px_2px_2px] hover:bg-gray-400/30 hover:shadow-[0px_0px_2px_2px] dark:bg-gray-400/20 dark:text-gray-100 dark:placeholder:text-gray-500"
+        className="focus-within:border-primary-600 mt-1 hover:border-primary-600 shadow-primary-300/40 flex w-full cursor-pointer items-center justify-center gap-2 rounded-[12px] border-2 border-transparent bg-gray-400/30 py-2 pl-4 text-xl outline-0 transition-all outline-none placeholder:text-xl placeholder:text-gray-600/50 focus-within:bg-gray-400/30 focus-within:shadow-[0px_0px_2px_2px] hover:bg-gray-400/30 hover:shadow-[0px_0px_2px_2px] dark:bg-gray-400/20 dark:text-gray-100 dark:placeholder:text-gray-500"
       >
         {/* Placeholder que força a escolha */}
         {defaultOption && (
-          <option value="" disabled>
+          <option value=""  className="dark:text-primary-170" disabled>
             {defaultOptionText}
           </option>
         )}
@@ -73,7 +73,7 @@ export default function FormInputSelect({
           <option
             key={option.id}
             value={option.id}
-            className="dark:text-primary-50"
+            className="dark:text-primary-170"
           >
             {option.nome}
           </option>
