+++ src/components/formSteps/createUser/FormStepsUser.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import React, { useActionState, useEffect } from "react";
-import { Form } from "../form";
+import { Form } from "../../form";
 import { toast } from "react-toastify";
 import { redirect } from "next/navigation";
-import RegisterNavigation from "../nav/registersNavigation/RegisterNavigation";
+import RegisterNavigation from "../../nav/registersNavigation/RegisterNavigation";
 import {
   Check,
   ChevronLeft,
@@ -15,7 +15,7 @@ import {
   UserCircle,
 } from "lucide-react";
 import { RoleData, PermissaoCompletaData } from "@/types/api/apiTypes";
-import PermissionsMenu from "./permissionsMenu/PermissionsMenu";
+import PermissionsMenu from "./permissionsForm/permissionsMenu/PermissionsMenu";
 import { postUser } from "@/actions/postUser";
 import { useGlobalContext } from "@/context/globalContext";
 
@@ -29,8 +29,15 @@ export default function FormStepsUser({
   permissoesData,
 }: FormStepsUser) {
   const [activeTab, setActiveTab] = React.useState("informacoes");
-  const [currentRoles, setCurrentRoles] = React.useState<RoleData[] | []>(rolesData);
+  const [currentRoles, setCurrentRoles] = React.useState<RoleData[] | []>(
+    rolesData,
+  );
   const { openGlobalModal, closeGlobalModal } = useGlobalContext();
+  const [senha, setSenha] = React.useState({
+    senha: "",
+    confirmar_senha: "",
+  });
+
   const [state, formAction] = useActionState(postUser, {
     errors: [],
     msg_success: "",
@@ -51,22 +58,24 @@ export default function FormStepsUser({
   }, [state]);
 
   return (
-    <div className="flex gap-5">
+    <div className="flex gap-5 ">
       <RegisterNavigation setActiveTab={setActiveTab} activeTab={activeTab} />
-      <Form.Root action={formAction}>
-        <div className={`${activeTab !== 'informacoes' ? 'hidden' : ''}`}>
-          <Form.Section title="Dados do usuario">
+      <Form.Root className="w-full" action={formAction}>
+        <div className={` ${activeTab !== "informacoes" ? "hidden" : ""}`}>
+          <Form.Section  title="Dados do usuario">
             <Form.InputText
               icon={UserCircle}
               inputId="nome"
               name="nome"
               label="Nome"
+              className="col-span-2"
             />
             <Form.InputText
               icon={Phone}
               inputId="contato"
               name="contato"
               label="Contato"
+              className="col-span-1"
             />
           </Form.Section>
           <Form.Section title="Acesso ao sistema">
@@ -76,6 +85,7 @@ export default function FormStepsUser({
               type="mail"
               name="email"
               label="Email"
+              className="col-span-2"
             />
             <Form.InputText
               icon={KeyRound}
@@ -83,6 +93,13 @@ export default function FormStepsUser({
               type="password"
               name="senha"
               label="Senha"
+              value={senha.senha}
+              onChange={(e) =>
+                setSenha({
+                  ...senha,
+                  senha: (e.target as HTMLInputElement).value,
+                })
+              }
             />
             <Form.InputText
               icon={KeyRound}
@@ -90,25 +107,37 @@ export default function FormStepsUser({
               type="password"
               name="senha"
               label="Confirmar Senha"
+              value={senha.confirmar_senha}
+              onChange={(e) =>
+                setSenha({
+                  ...senha,
+                  confirmar_senha: (e.target as HTMLInputElement).value,
+                })
+              }
             />
           </Form.Section>
           <Form.ButtonNext
             type="submit"
             onClick={(e) => {
               e.preventDefault();
-              setActiveTab("permissoes");
+              if (senha.senha !== senha.confirmar_senha) {
+                toast.error("As senhas não coincidem");
+              } else {
+                setActiveTab("permissoes");
+              }
             }}
             direction="Continuar"
             icon={ChevronRight}
           />
         </div>
-        <div className={`${activeTab !== 'permissoes' ? 'hidden' : ''}`}>
+        <div className={`${activeTab !== "permissoes" ? "hidden" : ""}`}>
           <Form.Section title="Perfil">
             <Form.InputSelect
               options={currentRoles}
               label="Tipo de usuario"
               id="tipo_usuario"
               name="tipo_usuario"
+              defaultOptionText="Selecione"
             />
           </Form.Section>
           <Form.Section title="Permissões">
@@ -124,11 +153,7 @@ export default function FormStepsUser({
               direction="Voltar"
               icon={ChevronLeft}
             />
-            <Form.ButtonNext
-              type="submit"
-              direction="Finalizar"
-              icon={Check}
-            />
+            <Form.ButtonNext type="submit" direction="Finalizar" icon={Check} />
           </div>
         </div>
       </Form.Root>
